Fix stateless TLS SIP connection
The request() function in re/src/sip/request.c fails with EINVAL
when sending stateless requests (for example ACKs) to a TLS server,
which makes SIP calls terminate.

This is because request() calls sip_send_conn() which always passes
a NULL host to sip_transp_send(). This gets passed down to conn_send()
in re/src/sip/transp.c and ultimately to tls_set_verify_server()
which fails with EINVAL.

This patch replaces sip_send_conn() with sip_transp_send(), so that
the host is passed down correctly.

Signed-off-by: Nicola Di Lieto <nicola.dilieto@gmail.com>